refactor(auth): move UserStatusType to its own file
